# src/swagger.json
{
  "swagger":"2.0"
  ,"info":{"description":"Lista Contatos API"
  ,"version":"1.0"
  ,"title":"Lista Contatos API"}
  ,"host":"localhost:18000"
  ,"basePath":"/"
...
  ,"schemes":["http"]
  ,"produces": ["application/json"],
  "paths": {
...
    "/contato/atualizar/{contatoId}": {
      "post": {
        "x-swagger-router-controller": "contatos",
        "operationId": "updateContato",
        "tags": ["Contatos"],
        "summary": "Update contato",
        "parameters": [ {
          "name": "contatoId",
          "in": "path",
          "type": "number",
          "required": true
        },
        {
          "name": "Contato",
          "in": "body",
          "schema" : {
            "$ref": "#/definitions/InsereContato"
          }
        }],
        "responses": {
          "200": {
            "description": "Get contato by id",
            "schema": {
              "$ref": "#/definitions/Contato"
            }
          }
        }
      }
    }
  },
  "definitions":{
    "Contato": {
      "properties": {
        "_id": {
          "type": "string",
          "example": 0
        },
        "nome": {
          "type": "string",
          "example":"Teste"
        },
        "email": {
          "type": "string",
          "example":"teste@teste.com"
        },
        "telefone": {
          "type": "string",
          "example":"(99) 99999-9999"
        }
      }
    },
    "InsereContato": {
      "properties": {
        "nome": {
          "type": "string",
          "example":"Teste"
        },
        "email": {
          "type": "string",
          "example":"teste@teste.com"
        },
        "telefone": {
          "type": "string",
          "example":"(99) 99999-9999"
        }
      }
    },
...
  }
}